--force flag added to mimic Tanzu scheduler plugin

+++ commands/delete-call-schedule.go
@@ -12,9 +12,12 @@ import (
 // cf delete-call-schedule CALL-NAME SCHEDULE-GUID
 func DeleteCallSchedule(services *core.Services, args []string) {
     var forceFlag bool
+    var promptFlag bool
 
 	flags := pflag.NewFlagSet("delete-call-schedule", pflag.ExitOnError)
     flags.BoolVarP(&forceFlag, "force", "f", false, "Force call schedule deletion without confirmation")
+    flags.BoolVarP(&promptFlag, "prompt", "p", false, "Allow call schedule deletion with confirmation")
+    flags.MarkHidden("prompt")
 	flags.Parse(args)
 	args = flags.Args()
     
@@ -37,6 +40,10 @@ func DeleteCallSchedule(services *core.Services, args []string) {
 		fmt.Printf("Could not find call named %s in space %s.\n", name, space.Name)
 		return
 	}
+
+    if promptFlag && !forceFlag && !services.UI.ConfirmDelete("call schedule", name + " " + scheduleGUID) {
+        return;
+    }
 
 	err = client.DeleteCallSchedule(services.Client, call, scheduleGUID)
 	if err != nil {
