+++ commands/delete-call.go
@@ -12,9 +12,12 @@ import (
 // cf delete-call CALL-NAME
 func DeleteCall(services *core.Services, args []string) {
     var forceFlag bool
+    var promptFlag bool
 
 	flags := pflag.NewFlagSet("delete-call", pflag.ExitOnError)
     flags.BoolVarP(&forceFlag, "force", "f", false, "Force call deletion without confirmation")
+    flags.BoolVarP(&promptFlag, "prompt", "p", false, "Allow call deletion with confirmation")
+    flags.MarkHidden("prompt")
 	flags.Parse(args)
 	args = flags.Args()
 
@@ -36,6 +39,10 @@ func DeleteCall(services *core.Services, args []string) {
 		fmt.Printf("Could not find call named %s in space %s.\n", name, space.Name)
 		return
 	}
+
+    if promptFlag && !forceFlag && !services.UI.ConfirmDeleteWithAssociations("call", name) {
+        return;
+    }
 
 	err = client.DeleteCall(services.Client, call)
 	if err != nil {
